chore: Move RiskColorMap to lib/analytics.ts
