+++ lib/analytics.ts
@@ -7,3 +7,31 @@ export const Factors = [
   'government-response',
   'exposure',
 ];
+
+export const RiskColorMap: {
+  [key: number]: {
+    backgroundColor: string;
+    indicatorColor: string;
+  };
+} = {
+  5: {
+    backgroundColor: '#d416057a',
+    indicatorColor: '#D41505',
+  },
+  4: {
+    backgroundColor: '#fb8b357a',
+    indicatorColor: '#FB8C35',
+  },
+  3: {
+    backgroundColor: '#ffee6e82',
+    indicatorColor: '#FFED6E',
+  },
+  2: {
+    backgroundColor: '#65a4bd77',
+    indicatorColor: '#65A4BD',
+  },
+  1: {
+    backgroundColor: '#4575b480',
+    indicatorColor: '#4575B4',
+  },
+};
